+++ packages/jest-jasmine2/src/queue_runner.js
@@ -34,7 +34,7 @@ export default function queueRunner(options: Options) {
     onCancel(resolve);
   });
 
-  const mapper = ({fn, timeout, timeoutError = new Error()}) => {
+  const mapper = ({fn, timeout, initError = new Error()}) => {
     let promise = new Promise(resolve => {
       const next = function(err) {
         if (err) {
@@ -69,11 +69,11 @@ export default function queueRunner(options: Options) {
       options.clearTimeout,
       options.setTimeout,
       () => {
-        timeoutError.message =
+        initError.message =
           'Timeout - Async callback was not invoked within the ' +
           timeoutMs +
           'ms timeout specified by jest.setTimeout.';
-        options.onException(timeoutError);
+        options.onException(initError);
       },
     );
   };